Update jquery.SPServices.js
Currently SPGetCurrentSite is called in case of a given webURL as well. If the page containing the script is not stored inside a SharePoint Site then the current site can't retrieved and on each call to SPServices the call is tried and failed again. Szenario: While developing I host my Markup locally and use SharePoint just as a generic service backend)

The local var thisSite is used inside the last else only. So I suggest to move the line to this last scope.