Fixed bug with TextFields
This find the comment nodes and only checks the comment nodes, not the full HTML.  I found a bug with the previous method in that it would return the incorrect `td` if a textfield had a value of `FieldName="ColumnName"`.  This new function also allows usage of the static name.